# contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Vehicle Inventory API
  description: API for managing the vehicle fleet inventory.
  version: 1.0.0
servers:
  - url: http://localhost:8000
    description: Local development server

paths:
  /health:
    get:
      summary: System Health Checkssss
      description: Verifies the active connection with MongoDB.
      responses:
        '200':
          description: System is healthy.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "ok"
                  database:
                    type: string
                    example: "connected"
        '500':
          description: System is unhealthy.

  /api/v1/vehicles:
    get:
      summary: List all vehicles
      description: Retrieve a paginated list of vehicles.
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 100
          description: Number of records to return.
        - in: query
          name: offset
          schema:
            type: integer
            default: 0
            minimum: 0
          description: Number of records to skip.
      responses:
        '200':
          description: List of vehicles.
          content:
            application/json:
              schema:
                type: object
                properties:
                  total:
                    type: integer
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Vehicle'
    post:
      summary: Create a new vehicle
      description: Register a new vehicle in the system.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VehicleInput'
      responses:
        '201':
          description: Vehicle created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vehicle'
        '400':
          description: Validation error.
        '409':
          description: Conflict (e.g., duplicate license plate or economic number).

  /api/v1/vehicles/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
        description: MongoDB Object ID of the vehicle.
    get:
      summary: Get a single vehicle
      description: Retrieve details of a specific vehicle by ID.
      responses:
        '200':
          description: Vehicle details.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vehicle'
        '404':
          description: Vehicle not found.
    put:
      summary: Update a vehicle
      description: Update details of an existing vehicle.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VehicleInput'
      responses:
        '200':
          description: Vehicle updated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vehicle'
        '404':
          description: Vehicle not found.
        '400':
          description: Validation error.
    delete:
      summary: Delete a vehicle
      description: Remove a vehicle from the system.
      responses:
        '204':
          description: Vehicle deleted successfully.
        '404':
          description: Vehicle not found.

components:
  schemas:
    VehicleInput:
      type: object
      required:
        - placa
        - numero_economico
        - marca
        - modelo
        - anno
        - tipo_vehiculo
        - capacidad_carga_kg
        - numero_serie
        - estado_vehiculo
        - poliza_seguro
        - vigencia_seguro
      properties:
        placa:
          type: string
          pattern: '^[A-Z0-9]{2,3}-[A-Z0-9]{3,5}$'
          example: "ABC-1234"
          description: Mexican license plate format.
        numero_economico:
          type: string
          example: "MTY-105"
        marca:
          type: string
          example: "Kenworth"
        modelo:
          type: string
          example: "T680"
        anno:
          type: integer
          minimum: 1990
          maximum: 2026
          example: 2024
        tipo_vehiculo:
          $ref: '#/components/schemas/VehicleType'
        capacidad_carga_kg:
          type: number
          format: float
          minimum: 0.01
          example: 35000.0
        numero_serie:
          type: string
          minLength: 17
          maxLength: 17
          example: "1NKDL4XDXMP123456"
        estado_vehiculo:
          $ref: '#/components/schemas/VehicleStatus'
        poliza_seguro:
          type: string
          example: "QUALITAS-998877"
        vigencia_seguro:
          type: string
          format: date
          example: "2027-01-29"
        kilometraje_actual:
          type: integer
          minimum: 0
          example: 15000
        tipo_combustible:
          $ref: '#/components/schemas/FuelType'

    Vehicle:
      allOf:
        - $ref: '#/components/schemas/VehicleInput'
        - type: object
          properties:
            _id:
              type: string
              description: MongoDB Object ID.
            fecha_alta:
              type: string
              format: date-time
              description: Creation timestamp.
            ultima_verificacion:
              type: string
              format: date-time
              description: Last verification timestamp.

    VehicleType:
      type: string
      enum:
        - TRACTOR_TRUCK
        - RIGID_TRUCK
        - TRAILER
        - DOLLY

    VehicleStatus:
      type: string
      enum:
        - ACTIVE
        - IN_MAINTENANCE
        - OUT_OF_SERVICE

    FuelType:
      type: string
      enum:
        - DIESEL
        - NATURAL_GAS
        - ELECTRIC